use actor to get actual types
