+++ src/lib/experimental/index.ts
@@ -2,13 +2,13 @@ import './experimental';
 import './globals';
 import '../../../type_tests';
 
+export * from '../stable/json';
 export * from '../stable/stable_structures/stable_b_tree_map';
 export * from '../stable/stable_structures/stable_json';
 export * from './candid';
 export * from './canister_methods';
 export { serialize } from './fetch';
 export * from './ic';
-export * from './json';
 export * from './server';
 export * from './system_types';
 export * from './threshold_wallet';
