+++ tests/end_to_end/candid_rpc/class_syntax/cross_canister_calls/benchmarks_canister1.md
@@ -2,19 +2,19 @@
 
 | Execution | Method Name      | Instructions | Cycles    | USD           |
 | --------- | ---------------- | ------------ | --------- | ------------- |
-| 0         | balance          | 2,297,063    | 1,508,825 | $0.0000020167 |
-| 1         | account          | 3,635,050    | 2,044,020 | $0.0000027321 |
-| 2         | balance          | 2,217,475    | 1,476,990 | $0.0000019742 |
-| 3         | account          | 3,614,703    | 2,035,881 | $0.0000027212 |
-| 4         | accounts         | 1,654,333    | 1,251,733 | $0.0000016731 |
-| 5         | transfer         | 3,564,397    | 2,015,759 | $0.0000026943 |
-| 6         | balance          | 2,215,122    | 1,476,049 | $0.0000019729 |
-| 7         | account          | 3,613,781    | 2,035,512 | $0.0000027207 |
-| 8         | balance          | 2,208,167    | 1,473,267 | $0.0000019692 |
-| 9         | account          | 3,607,272    | 2,032,909 | $0.0000027172 |
-| 10        | accounts         | 1,651,338    | 1,250,535 | $0.0000016715 |
-| 11        | trap             | 1,623,830    | 1,239,532 | $0.0000016568 |
-| 12        | sendNotification | 2,648,666    | 1,649,466 | $0.0000022047 |
+| 0         | balance          | 2,297,155    | 1,508,862 | $0.0000020168 |
+| 1         | account          | 3,635,303    | 2,044,121 | $0.0000027322 |
+| 2         | balance          | 2,217,498    | 1,476,999 | $0.0000019742 |
+| 3         | account          | 3,614,819    | 2,035,927 | $0.0000027212 |
+| 4         | accounts         | 1,654,437    | 1,251,774 | $0.0000016731 |
+| 5         | transfer         | 3,564,397    | 2,015,758 | $0.0000026943 |
+| 6         | balance          | 2,215,030    | 1,476,012 | $0.0000019729 |
+| 7         | account          | 3,613,862    | 2,035,544 | $0.0000027207 |
+| 8         | balance          | 2,208,190    | 1,473,276 | $0.0000019692 |
+| 9         | account          | 3,607,099    | 2,032,839 | $0.0000027171 |
+| 10        | accounts         | 1,651,246    | 1,250,498 | $0.0000016714 |
+| 11        | trap             | 1,623,767    | 1,239,506 | $0.0000016567 |
+| 12        | sendNotification | 2,648,598    | 1,649,439 | $0.0000022047 |
 
 ---
