+++ tests/end_to_end/candid_rpc/class_syntax/cross_canister_calls/benchmarks_canister2.md
@@ -2,8 +2,8 @@
 
 | Execution | Method Name         | Instructions | Cycles    | USD           |
 | --------- | ------------------- | ------------ | --------- | ------------- |
-| 0         | transfer            | 2,156,919    | 1,452,768 | $0.0000019418 |
-| 1         | receiveNotification | 1,384,872    | 1,143,949 | $0.0000015290 |
+| 0         | transfer            | 2,156,919    | 1,452,767 | $0.0000019418 |
+| 1         | receiveNotification | 1,384,872    | 1,143,948 | $0.0000015290 |
 
 ---
